requirements.txt
django==3.2.18
djangorestframework==3.12.1
psycopg2==2.8
